+++ citrixadc/resource_citrixadc_auditsyslogaction.go
@@ -172,17 +172,17 @@ func createAuditsyslogactionFunc(d *schema.ResourceData, meta interface{}) error
 		Dateformat:           d.Get("dateformat").(string),
 		Dns:                  d.Get("dns").(string),
 		Domainresolvenow:     d.Get("domainresolvenow").(bool),
-		Domainresolveretry:   int32(d.Get("domainresolveretry").(int)),
+		Domainresolveretry:   d.Get("domainresolveretry").(int),
 		Lbvservername:        d.Get("lbvservername").(string),
 		Logfacility:          d.Get("logfacility").(string),
 		Loglevel:             toStringList(loglevelValue(d)),
 		Lsn:                  d.Get("lsn").(string),
-		Maxlogdatasizetohold: uint32(d.Get("maxlogdatasizetohold").(int)),
+		Maxlogdatasizetohold: d.Get("maxlogdatasizetohold").(int),
 		Name:                 d.Get("name").(string),
 		Netprofile:           d.Get("netprofile").(string),
 		Serverdomainname:     d.Get("serverdomainname").(string),
 		Serverip:             d.Get("serverip").(string),
-		Serverport:           int32(d.Get("serverport").(int)),
+		Serverport:           d.Get("serverport").(int),
 		Sslinterception:      d.Get("sslinterception").(string),
 		Subscriberlog:        d.Get("subscriberlog").(string),
 		Tcp:                  d.Get("tcp").(string),
@@ -297,7 +297,7 @@ func updateAuditsyslogactionFunc(d *schema.ResourceData, meta interface{}) error
 	}
 	if d.HasChange("domainresolveretry") {
 		log.Printf("[DEBUG]  citrixadc-provider: Domainresolveretry has changed for auditsyslogaction %s, starting update", auditsyslogactionName)
-		auditsyslogaction.Domainresolveretry = int32(d.Get("domainresolveretry").(int))
+		auditsyslogaction.Domainresolveretry = d.Get("domainresolveretry").(int)
 		hasChange = true
 	}
 	if d.HasChange("lbvservername") {
@@ -322,7 +322,7 @@ func updateAuditsyslogactionFunc(d *schema.ResourceData, meta interface{}) error
 	}
 	if d.HasChange("maxlogdatasizetohold") {
 		log.Printf("[DEBUG]  citrixadc-provider: Maxlogdatasizetohold has changed for auditsyslogaction %s, starting update", auditsyslogactionName)
-		auditsyslogaction.Maxlogdatasizetohold = uint32(d.Get("maxlogdatasizetohold").(int))
+		auditsyslogaction.Maxlogdatasizetohold = d.Get("maxlogdatasizetohold").(int)
 		hasChange = true
 	}
 	if d.HasChange("name") {
@@ -347,7 +347,7 @@ func updateAuditsyslogactionFunc(d *schema.ResourceData, meta interface{}) error
 	}
 	if d.HasChange("serverport") {
 		log.Printf("[DEBUG]  citrixadc-provider: Serverport has changed for auditsyslogaction %s, starting update", auditsyslogactionName)
-		auditsyslogaction.Serverport = int32(d.Get("serverport").(int))
+		auditsyslogaction.Serverport = d.Get("serverport").(int)
 		hasChange = true
 	}
 	if d.HasChange("sslinterception") {
